Fix combined copy on startup and create on empty
If both options are set to true and a database file does not exist a new
database file is created and all data saved during the kbase session
will persist in the file. If a database file already exists the contents
are copied to a temporary database and newly saved data will not
persist.